+++ fili-core/src/main/java/com/yahoo/bard/webservice/druid/model/aggregation/LongMaxAggregation.java
@@ -2,6 +2,8 @@
 // Licensed under the terms of the Apache license. Please see LICENSE.md file distributed with this work for terms.
 package com.yahoo.bard.webservice.druid.model.aggregation;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
+
 /**
  * Aggregation for the maximum of longs.
  */
@@ -30,5 +32,10 @@ public class LongMaxAggregation extends Aggregation {
     @Override
     public Aggregation withFieldName(String fieldName) {
         return new LongMaxAggregation(getName(), fieldName);
+    }
+
+    @JsonIgnore
+    public boolean isFloatingPoint() {
+        return false;
     }
 }